move config.ts to config/
